feat: Add Redux store and cart functionality
The code changes include adding Redux store configuration and cart functionality to the application. The `Provider` component from the `react-redux` library is imported and used to wrap the `ThemeProvider` and `App` components in the `index.js` file. The `configureStore` function from `@reduxjs/toolkit` is used to create the store with the `cartReducer` as the root reducer. The `cartReducer` handles actions related to the cart, such as adding items, removing items, and updating the cart count. The `cart` state is also accessed in the `Navbar` component to display the number of items in the cart.